+++ src/pages/BO/login/index.ts
@@ -1,5 +1,5 @@
 import type {LoginPageInterface} from '@interfaces/BO/login';
-import testContext from '@utils/testContext';
+import testContext from '@utils/test';
 import semver from 'semver';
 
 const psVersion = testContext.getPSVersion();
