+++ src/pages/FO/classic/home/index.ts
@@ -1,5 +1,5 @@
 import type {FoHomePageInterface} from '@interfaces/FO/home';
-import testContext from '@utils/testContext';
+import testContext from '@utils/test';
 import semver from 'semver';
 
 const psVersion = testContext.getPSVersion();
